requirements.txt
streamlit
pandas
sentence-transformers
chromadb
semantic-router==0.0.20
python-dotenv
openai
